+++ lib/bioruby.rb
@@ -1,8 +1,8 @@
 # frozen_string_literal: true
 
 require_relative 'bioruby/mcp/server/version'
-require_relative 'bioruby/mcp/server/core'
 require_relative 'bioruby/mcp/server/kegg_tools'
+require_relative 'bioruby/mcp/server/core'
 
 module BioRuby
   module MCP